requirements.txt
Flask==3.0.3
Jinja2==3.1.3
openai==1.25.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==1.26.6
